refactor(middleware): reorganize directory structure and add RequestContextMiddleware
- Moved middleware into more specific namespaces (e.g., `Auth`, `Localization`, `Context`).
- Introduced `AbstractContextMiddleware` for shared context handling logic.
- Added `RequestContextMiddleware` to streamline context initialization.
