+++ app/Http/Middleware/Auth/AuthenticateOrShowMessageMiddleware.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Http\Middleware;
+namespace App\Http\Middleware\Auth;
 
 use Closure;
 use Illuminate\Http\Request;
