+++ app/Http/Middleware/Auth/EnsureEmailIsVerifiedMiddleware.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Middleware;
+namespace App\Http\Middleware\Auth;
 
 use App\Models\User;
 use Closure;
